Add animation on delete
This required wrapping each row in a div that can be animated to wipe in/out the row visually. The wrapper has no margin or padding, so animated cleanly. The contents of the row are each in a separte grid element now. They line up since their contents are all the same.